voice_pipeline/requirements.txt
# LiveKit Voice Agent Dependencies
livekit>=1.0.0
livekit-agents[openai,silero,assemblyai,turn-detector,deepgram,google,anthropic,cartesia,elevenlabs]>=1.2
livekit-plugins-noise-cancellation

# Core dependencies
python-dotenv>=1.0.0
certifi>=2024.2.2
av>=10.0.0

# HTTP client for backend calls
httpx>=0.24.0

# Web framework for session API
fastapi>=0.110.0
uvicorn>=0.30.0

# Configuration
pyyaml>=6.0

# Metrics and logging
wandb>=0.15.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Modal deployment (optional, for cloud deployment)
modal>=0.55.0
